added v6

+++ src/main/java/dev/tsvinc/r2p/PaymentRequest.java
@@ -2,4 +2,6 @@ package dev.tsvinc.r2p;
 
 import jakarta.validation.constraints.DecimalMin;
 
-public record PaymentRequest(@DecimalMin(value = "0.01") Double requestedAmount) {}
+import java.math.BigDecimal;
+
+public record PaymentRequest(@DecimalMin(value = "0.01") BigDecimal requestedAmount) {}
